Move results from utils
