+++ model_api/python/model_api/models/visual_prompting.py
@@ -12,9 +12,10 @@ from typing import Any, NamedTuple
 import cv2
 import numpy as np
 
-from model_api.models import SAMDecoder, SAMImageEncoder
-from model_api.models.utils import (
+from model_api.models import (
     PredictedMask,
+    SAMDecoder,
+    SAMImageEncoder,
     VisualPromptingResult,
     ZSLVisualPromptingResult,
 )
