+++ model_api/python/model_api/models/yolo.py
@@ -11,8 +11,9 @@ import numpy as np
 from model_api.adapters.utils import INTERPOLATION_TYPES, resize_image_ocv
 
 from .detection_model import DetectionModel
+from .result_types import Detection, DetectionResult
 from .types import BooleanValue, ListValue, NumericalValue
-from .utils import Detection, DetectionResult, clip_detections, multiclass_nms, nms
+from .utils import clip_detections, multiclass_nms, nms
 
 DetectionBox = namedtuple("DetectionBox", ["x", "y", "w", "h"])
